tweak how gsl::span and std::span are exposed
